+++ ta4j-core/src/test/java/org/ta4j/core/criteria/LinearTransactionCostCriterionTest.java
@@ -70,17 +70,17 @@ public class LinearTransactionCostCriterionTest extends AbstractCriterionTest {
         TradingRecord tradingRecord = new BaseTradingRecord();
         Num criterion;
 
-        tradingRecord.operate(0);
-        tradingRecord.operate(1);
+        tradingRecord.trade(0);
+        tradingRecord.trade(1);
         criterion = getCriterion(1000d, 0.005, 0.2).calculate(series, tradingRecord);
         assertNumEquals(12.861, criterion);
 
-        tradingRecord.operate(2);
-        tradingRecord.operate(3);
+        tradingRecord.trade(2);
+        tradingRecord.trade(3);
         criterion = getCriterion(1000d, 0.005, 0.2).calculate(series, tradingRecord);
         assertNumEquals(24.3759, criterion);
 
-        tradingRecord.operate(5);
+        tradingRecord.trade(5);
         criterion = getCriterion(1000d, 0.005, 0.2).calculate(series, tradingRecord);
         assertNumEquals(28.2488, criterion);
     }
@@ -91,17 +91,17 @@ public class LinearTransactionCostCriterionTest extends AbstractCriterionTest {
         TradingRecord tradingRecord = new BaseTradingRecord();
         Num criterion;
 
-        tradingRecord.operate(0);
-        tradingRecord.operate(1);
+        tradingRecord.trade(0);
+        tradingRecord.trade(1);
         criterion = getCriterion(1000d, 0d, 1.3d).calculate(series, tradingRecord);
         assertNumEquals(2.6d, criterion);
 
-        tradingRecord.operate(2);
-        tradingRecord.operate(3);
+        tradingRecord.trade(2);
+        tradingRecord.trade(3);
         criterion = getCriterion(1000d, 0d, 1.3d).calculate(series, tradingRecord);
         assertNumEquals(5.2d, criterion);
 
-        tradingRecord.operate(0);
+        tradingRecord.trade(0);
         criterion = getCriterion(1000d, 0d, 1.3d).calculate(series, tradingRecord);
         assertNumEquals(6.5d, criterion);
     }
@@ -115,15 +115,15 @@ public class LinearTransactionCostCriterionTest extends AbstractCriterionTest {
         criterion = getCriterion(1000d, 0d, 0.75d).calculate(series, position);
         assertNumEquals(0d, criterion);
 
-        position.operate(1);
+        position.trade(1);
         criterion = getCriterion(1000d, 0d, 0.75d).calculate(series, position);
         assertNumEquals(0.75d, criterion);
 
-        position.operate(3);
+        position.trade(3);
         criterion = getCriterion(1000d, 0d, 0.75d).calculate(series, position);
         assertNumEquals(1.5d, criterion);
 
-        position.operate(4);
+        position.trade(4);
         criterion = getCriterion(1000d, 0d, 0.75d).calculate(series, position);
         assertNumEquals(1.5d, criterion);
     }